moved style.css to stylesheets folder
